commit
added 3 controllers for the mvc so the other sections of the website will operate, additional css was added to the website, added a dashboard to the homepage, with the buttons in the middle that takes you to the other sections of the website. further improvements were made to the submit, management and track pages.